register login

+++ app/Http/Controllers/TasksController.php
@@ -13,12 +13,21 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+      
     public function index()
-    {
-        $tasks=Task::all();
-        
-        return view('tasks.index',['tasks'=>$tasks,]);
+    {   
+        if(\Auth::check()){
+            $user=\Auth::user();
+            
+            $tasks=$user->tasks()->orderBy('created_at','desc')->paginate(3);
+            return view('tasks.index',['tasks'=>$tasks,'user'=>$user]);
+        }
+        else{
+            return view('welcome');
+        }
     }
+    
+    
 
     /**
      * Show the form for creating a new resource.
@@ -41,14 +50,17 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'status'=>'required|max:10',
-            'content'=>'required|max:255',]);
-        $task=new Task;
-        $task->status=$request->status;
-        $task->content = $request->content;
-        $task->save();
-        
-        return redirect('/');
+                'status'=>'required|max:255',
+                'content'=>'required|max:255',
+            ]);
+            
+        $request->user()->tasks()->create([
+                'status'=>$request->status,
+                'content'=>$request->content,
+                
+            ]);
+            
+        return redirect('/');    
     }
 
     /**
@@ -60,8 +72,9 @@ class TasksController extends Controller
     public function show($id)
     {
         $task=Task::findOrFail($id);
+        $user=\Auth::user();
         
-        return view('tasks.show',['task'=>$task,]);
+        return view('tasks.show',['task'=>$task,'user'=>$user,]);
     }
 
     /**
@@ -73,8 +86,9 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task=Task::findOrFail($id);
+        $user=\Auth::user();
         
-        return view('tasks.edit',['task'=>$task,]);
+        return view('tasks.edit',['task'=>$task,'user'=>$user,]);
     }
 
     /**
@@ -84,18 +98,28 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request)
     {
         $request->validate([
             'status'=>'required|max:10',
             'content'=>'required|max:255',
             ]);
-        $task=Task::findOrFail($id);
+        
+         $request->user()->tasks()->update([
+                'status'=>$request->status,
+                'content'=>$request->content,]);
+                
+        return redirect('/');
+        
+            
+            
+       /*$task=Task::findOrFail($id);
         $task->status=$request->status;
         $task->content=$request->content;
+        $task->user_id=$request->user_id;
         $task->save();
         
-        return redirect('/');
+        return redirect('/');*/
     }
 
     /**
@@ -106,9 +130,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task=Task::findOrFail($id);
-        $task->delete();
+        $task=\App\Task::findOrFail($id);
         
+        if(\Auth::id()===$task->user_id){
+        $task->delete();
+        }
         return redirect('/');
     }
 }
